[Shaman] Refactor/write Fury of Air
Now is player-based aoe damage instead of "a dot". Also adds support for
canceling Fury of Air by executing the action when Fury of Air is
running. Canceling incurs no GCD, and costs maelstrom proportional to
the current elapsed tick time.

Default APL updated to cancel when the number of enemies condition
fails, also updated to use spell_targets for the Fury of Air damage
spell so that when distance targeting is used, proper target counting is
enforced.