[Death Knight] Adjust Harrowing Decay
(should've been done before bfa launch)
Harrowing Decay is weird: it's a copy of the old unholy T21 2P bonus that applied a dot equal to a % of the damage dealt by Death Coil. This mechanic worked like Ignite and the new damage was added to the current dot if the effect was refreshed. This works well for a dot that is based on damage dealt because it doesn't have to scale with anything to avoid double dipping, and the duration was capped at current tick remaining + full duration.
Blizzard made an effort to properly copy this set bonus with Harrowing Decay, so the damage is added to the current dot just like before BUT this dot can now crit and scales with versatility, spec aura and mastery as well. On top of that the duration now pandemics.